fetch/stale-while-revalidate/stale-script.tentative.html

<!DOCTYPE html>
<!---
Tentative test against:
https://github.com/whatwg/fetch/pull/853
-->
<meta charset="utf-8">
<title>Tests Stale While Revalidate works for scripts</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
var last_modified;
var last_modified_count = 0;
var seen_revalidation = false;

// The script will call report via a uniquely generated ID on the subresource.
// If it is a cache hit the ID will be the same and the test will pass.
function report(mod, revalidation) {
  if (!last_modified) {
    last_modified = mod;
    last_modified_count = 1;
  } else if (last_modified == mod) {
    last_modified_count++;
  }

  if (revalidation == "True") {
    seen_revalidation = true;
  }
}

async_test(t => {
  window.onload = t.step_func(() => {
    step_timeout(() => {
      var script = document.createElement("script");
      script.src = "stale-script.py";
      document.body.appendChild(script);
      script.onload = t.step_func(() => {
          assert_equals(last_modified_count, 2, "last modified");
          assert_false(seen_revalidation);
          var checkResult = () => {
            var script = document.createElement("script");
            script.src = "stale-script.py";
            script.onload = t.step_func(() => {
              if (seen_revalidation) {
                  t.done();
              } else {
                t.step_timeout(checkResult, 25);
              }
            });
            document.body.appendChild(script);
          };
          t.step_timeout(checkResult, 25);
      });
    }, 0);
  });
}, 'Cache returns stale resource');

var script = document.createElement("script");
script.src = "stale-script.py";
document.body.appendChild(script);
</script>
</body>
